requirements.txt
torch
transformers
discord.py
Pillow
requests
numpy